+++ tests/epyccel/test_epyccel_classes.py
@@ -28,3 +28,15 @@ def test_class_return(modnew):
 
     b = modnew.get_B(3.0)
     assert isinstance(b, modnew.B)
+
+
+def test_class_argument(modnew):
+    b = modnew.get_B(3.0)
+    assert isinstance(b, modnew.B)
+
+    x = modnew.get_x_from_B(b)
+    assert x == 3.0
+    x = modnew.get_an_x_from_B(b)
+    assert x == 3.0
+    x = modnew.get_an_x_from_B()
+    assert x == -2.0